+++ python/setup.cfg
@@ -1,0 +1,2 @@
+[pytest]
+norecursedirs=tests/helpers
