+++ python/tests/api/logger/test_rolling.py
@@ -2,7 +2,6 @@ import functools
 import math
 import os
 
-# import time
 from datetime import datetime, timezone
 from os import listdir
 from os.path import isfile
@@ -11,6 +10,7 @@ from unittest.mock import MagicMock
 
 import pandas as pd
 import pytest
+import utils as u
 
 import whylogs as why
 from whylogs.api.logger.rolling import Scheduler
@@ -22,48 +22,11 @@ from whylogs.core.errors import BadConfigError
 from whylogs.core.schema import DatasetSchema
 from whylogs.core.segmentation_partition import segment_on_column
 
-# A sequence of timers get instantiated, so we need global state for them
-# to simulate universal time.
-
-_FAKE_TIMES: List[float] = []  # list of times (seconds) where the FakeTimer checks if it should execute
-_TIME_INDEX: int = 0  # where are we in the lsit of times
-
-
-class FakeTimer:
-    """
-    Fake for threading.Timer, not threaded. Evaluates elapsed duration
-    at times in the _FAKE_TIME list.
-    """
-
-    def __init__(self, interval: float, fn: Callable) -> None:
-        self._interval = interval
-        self._fn = fn
-        self._prev_start_time = 0.0
-        if _TIME_INDEX < len(_FAKE_TIMES):
-            self._prev_start_time = _FAKE_TIMES[_TIME_INDEX]
-
-    def start(self) -> None:
-        global _TIME_INDEX
-        _TIME_INDEX += 1
-        if _TIME_INDEX >= len(_FAKE_TIMES):
-            return
-
-        now = _FAKE_TIMES[_TIME_INDEX]
-        if (now - self._prev_start_time) >= self._interval:
-            self._fn()
-            self._prev_start_time = now
-        else:
-            self.start()
-
-    def cancel(self) -> None:
-        global _TIME_INDEX
-        _TIME_INDEX = len(_FAKE_TIMES) + 1
-
 
 class TimerContext:
     def __enter__(self) -> "TimerContext":
         self._prev_timer = Scheduler._TIMER
-        Scheduler._TIMER = FakeTimer
+        Scheduler._TIMER = u.FakeTimer
         return self
 
     def __exit__(self, exc_type, exc_value, exc_tb) -> None:
@@ -126,38 +89,30 @@ def test_rolling(tmp_path: Any) -> None:
     d = {"col1": [1, 2], "col2": [3.0, 4.0], "col3": ["a", "b"]}
     df = pd.DataFrame(data=d)
 
-    global _FAKE_TIMES, _TIME_INDEX
-    _FAKE_TIMES = [0, 0.1]
-    _TIME_INDEX = 0
-
-    # start = time.time()
+    u._FAKE_TIMES = [0, 0.1]  # time list must exist at logger instantiation, but don't fire yet
+    u._TIME_INDEX = 0
     with why.logger(mode="rolling", interval=1, when="S", base_name="test_base_name") as logger:
         logger.append_writer(writer=TestWriter(base_dir=tmp_path))
         logger.log(df)
-        _FAKE_TIMES = [x / 10.0 for x in range(16)]
-        _TIME_INDEX = 0
+        u._FAKE_TIMES = [x / 10.0 for x in range(16)]  # run for 1.5s, might fire 1 or 2 times depending on initial_run_after
+        u._TIME_INDEX = 0
         logger._scheduler._timer.start()
-        # time.sleep(1.5)
 
-        # Note that the number of files generated is depend on the elapsed amount
-        elapsed = _FAKE_TIMES[-1]
-        # elapsed = time.time() - start
-
+        # Note that the number of files generated is depend on the elapsed time
+        elapsed = u._FAKE_TIMES[-1]
         assert math.floor(elapsed) <= count_files(tmp_path) <= math.ceil(elapsed)
 
-        _FAKE_TIMES = [1.5 + x / 10.0 for x in range(17)]
-        _TIME_INDEX = 0
+        u._FAKE_TIMES = [1.5 + x / 10.0 for x in range(17)]  # run another 1.5s, fires 1 or 2 times depending on alignment
+        u._TIME_INDEX = 0
         logger.log(df)
         logger._scheduler._timer.start()
-        # time.sleep(1.5)
-        elapsed = _FAKE_TIMES[-1]
-        # elapsed = time.time() - start
+
+        elapsed = u._FAKE_TIMES[-1]
         assert math.floor(elapsed) <= count_files(tmp_path) <= math.ceil(elapsed)
 
         logger.log(df)
 
-    elapsed = _FAKE_TIMES[-1]
-    # elapsed = time.time() - start
+    elapsed = u._FAKE_TIMES[-1]
     assert math.floor(elapsed) <= count_files(tmp_path) <= math.ceil(elapsed)
 
 
@@ -166,19 +121,17 @@ def test_rolling_with_callback(tmp_path: Any) -> None:
     rolling_callback = MagicMock()
     messages = [{"col1": i, "col2": i * i * 1.2, "col3": "a"} for i in range(10)]
 
-    global _FAKE_TIMES, _TIME_INDEX
-    _FAKE_TIMES = [0, 0]
-    _TIME_INDEX = 0
+    u._FAKE_TIMES = [0, 0]
+    u._TIME_INDEX = 0
     rolling_logger = why.logger(
         mode="rolling", interval=1, when="S", base_name="test_base_name", callback=rolling_callback
     )
     rolling_logger.append_writer("local", base_dir=tmp_path)
     # process the 10 input messages, and wait a second to allow the rolling logger to hit an interval
     map(rolling_logger.log, messages)
-    _FAKE_TIMES = [0, 1]
-    _TIME_INDEX = 0
+    u._FAKE_TIMES = [0, 1]
+    u._TIME_INDEX = 0
     rolling_logger._scheduler._timer.start()
-    # time.sleep(1)
 
     # without an explicit calls to rolling_logger.flush we expect that the elapsed time
     # is greater than 1s interval, so the callback should have been triggered at least once
@@ -198,16 +151,14 @@ def test_rolling_skip_empty(tmp_path: Any) -> None:
     d = {"col1": [1, 2], "col2": [3.0, 4.0], "col3": ["a", "b"]}
     df = pd.DataFrame(data=d)
 
-    global _FAKE_TIMES, _TIME_INDEX
-    _FAKE_TIMES = [0, 0]
-    _TIME_INDEX = 0
+    u._FAKE_TIMES = [0, 0]
+    u._TIME_INDEX = 0
     with why.logger(mode="rolling", interval=1, when="S", base_name="test_base_name", skip_empty=True) as logger:
         logger.append_writer(writer=TestWriter(base_dir=tmp_path))
         logger.log(df)
-        _FAKE_TIMES = [0, 1, 2.1]
-        _TIME_INDEX = 0
+        u._FAKE_TIMES = [0, 1, 2.1]
+        u._TIME_INDEX = 0
         logger._scheduler._timer.start()
-        # time.sleep(2.1)
 
         # Note we sleep for over 2 seconds and have rolling interval of 1 second, so we should see at
         # least two elapsed intervals, but have only one file with skip_empty true
@@ -215,8 +166,8 @@ def test_rolling_skip_empty(tmp_path: Any) -> None:
 
         # log one more time, but don't wait and rely on the lifecycle of exiting the above with clause
         # to trigger at least one more flush with some new data
-        _FAKE_TIMES = [2.1, 2.2]
-        _TIME_INDEX = 0
+        u._FAKE_TIMES = [2.1, 2.2]
+        u._TIME_INDEX = 0
         logger.log(df)
         logger._scheduler._timer.start()
     assert count_files(tmp_path) == 2
@@ -247,12 +198,10 @@ def test_rolling_with_local_store_writes() -> None:
 
     with why.logger(mode="rolling", interval=1, when="S", base_name="test_base_name", skip_empty=True) as logger:
         logger.append_store(store=store)
-        global _FAKE_TIMES, _TIME_INDEX
-        _FAKE_TIMES = [0, 1.1]
-        _TIME_INDEX = 0
+        u._FAKE_TIMES = [0, 1.1]  # sleep 1s so it fires once
+        u._TIME_INDEX = 0
         logger.log(df)
         logger._scheduler._timer.start()
-        # time.sleep(1)
         assert len(store.list()) == 1
 
         query = DatasetIdQuery(dataset_id="test_base_name")
